+++ src/extensions/command/commandext_hooks.cpp
@@ -212,6 +212,7 @@ void Init_Vinifera_Commands()
     Commands.Add(new HealthFilterCommandClass);
     Commands.Add(new HealthFilterAddNextCommandClass);
     Commands.Add(new BeaconPlacementCommandClass);
+    Commands.Add(new ToggleDebugOverlayCommandClass);
 
     /**
      *  Initialize hotkeys for sidebar tabs.